fix:[scope:修复A区写入bug]修复写入bug,发现新bug 1.文件无法通过xmodem协议传输成功 2.推测可能是CRC校验函数的问题

+++ (B_Partition)/Project/BSP/fmc.c
@@ -1,7 +1,7 @@
 #include "fmc.h"
 #include "uart.h"
 
-void gd32_erase_flash(uint16_t start,uint16_t num)
+void gd32_erase_flash(uint32_t start,uint16_t num)
 {
 	
 	/* 解锁flash */
